Fix component error.
