+++ LLDebugTool/Core/Component/Hierarchy/UserInterface/LLHierarchyViewController.m
@@ -31,11 +31,11 @@
 #import "LLWindowManager.h"
 #import "LLThemeManager.h"
 #import "LLFactory.h"
+#import "LLConfig.h"
 #import "LLConst.h"
 #import "LLTool.h"
 
 #import "UIViewController+LL_Utils.h"
-#import "LLConfig+Hierarchy.h"
 #import "NSObject+LL_Utils.h"
 #import "UIView+LL_Utils.h"
