--- LLDebugTool/Core/Component/Log/Function/LLConfig+Log.h
@@ -1,69 +1,0 @@
-//
-//  LLConfig+Log.h
-//
-//  Copyright (c) 2018 LLDebugTool Software Foundation (https://github.com/HDB-Li/LLDebugTool)
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files (the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions:
-//
-//  The above copyright notice and this permission notice shall be included in all
-//  copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-//  SOFTWARE.
-
-#import "LLConfig.h"
-
-NS_ASSUME_NONNULL_BEGIN
-
-/**
- Log style for [LLDebugTool logInFile...]. Customize the log you want.
- 
- - LLConfigLogDetail: Show all detail info. Contain event, file, line, func, date and desc.
- - LLConfigLogFileFuncDesc : Show with event, file, func and desc.
- - LLConfigLogFileDesc : Show with event, file and desc.
- - LLConfigLogNormal: Show as system NSLog
- - LLConfigLogNone: Don't show anything.
- */
-typedef NS_ENUM(NSUInteger, LLConfigLogStyle) {
-    LLConfigLogDetail,
-    LLConfigLogFileFuncDesc,
-    LLConfigLogFileDesc,
-    LLConfigLogNormal,
-    LLConfigLogNone,
-};
-
-/**
- Log Level. It can be used for filter.
- 
- - LLConfigLogLevelDefault: Use to save message or note.
- - LLConfigLogLevelAlert: Use to save alert message.
- - LLConfigLogLevelWarning: Use to save warning message.
- - LLConfigLogLevelError: Use to save error message.
- */
-typedef NS_ENUM(NSUInteger, LLConfigLogLevel) {
-    LLConfigLogLevelDefault,
-    LLConfigLogLevelAlert,
-    LLConfigLogLevelWarning,
-    LLConfigLogLevelError,
-};
-
-@interface LLConfig (Log)
-
-/**
- Customize the log style. Default is LLConfigLogDetail.
- */
-@property (nonatomic, assign) LLConfigLogStyle logStyle;
-
-@end
-
-NS_ASSUME_NONNULL_END
