+++ LLDebugTool/Core/Component/Log/Function/LLLogHelper.h
@@ -23,7 +23,7 @@
 
 #import <Foundation/Foundation.h>
 
-#import "LLConfig+Log.h"
+#import "LLConfig.h"
 
 NS_ASSUME_NONNULL_BEGIN
